requirements.txt
numpy==1.24.4
flask
flask-cors
torch==2.1.0
torchvision==0.16.0
pillow
requests
transformers
diffusers
accelerate
imagehash
git+https://github.com/openai/CLIP.git#egg=clip
